requirements.txt
fastapi
uvicorn
openai
pydantic
google-generativeai
python-dotenv
langgraph
langchain
langchain-google-genai